fix: explicitly set the additional types to be loaded
If this is not set here, IDE tools such as VS Code will load all
typescript definitions from all node_modules in the tree.

Since both jest and mocha are in the root node_modules folder and each
of them have a declaration for the describe and it function used in our
tests, it doesn't know which one to take and intellisense is wrong.

This is a common issue in a monorepo using workspaces where all
dependencies of all packages are hoisted in the root folder.