Fix List shows empty page too often
The default empty list page, which is an invite to create new elements, appears even when some elements already exist.

Here is a repro that can be tested in the simple example:

1. Go to the post list (in a large screen to show the datagrid version)
2. Select all rows using the top left checkbox
3. Hit the bulk delete button

The empty page shows up, even though there are remaining posts (in the following page).

This bug was introduced by #8945, which fixed a bug with partial pagination.

The condition for displaying the empty page is wrong:

https://github.com/marmelab/react-admin/blob/e6129b2d1aded97f55bf40bf1d5602cc42a03113/packages/ra-ui-materialui/src/list/ListView.tsx#L60-L64