Fix `ra-data-fakerest` removes data when a form has an input `disabled`
## Problem

React-hook-form sets the value of `disabled` inputs to `undefined`. For most data providers, this isn't a problem (except the disabled field won't be sent in the update).

But FakeRest is local, so it receives the `undefined` value, and uses it to override the existing value. In effect, this erases the value.

This can be seen e.g. in the simple example, by changing the first input of PostEdit to `<TexuInput source="id" disabled" />`

## Solution

Remove `undefined` values from the edit payload.