Fix FilterButton checked status
## Problem

In the new design of the FilterButton dropdown, the checkbox for a filter is checked is the filter has a value. This isn't intuitive: if a user clicks on the item to have a filter input appear, they expect the checkbox to be displayed next time they open the dropdown.

## Solution

Check the checkbox if the filter is displayed, not applied.